refactor: modularize template configs
