+++ frontends/nextjs/src/lib/nerd-mode-ide/templates/configs/advanced.ts
@@ -1,0 +1,3 @@
+import type { PackageTemplateConfig } from '../../types'
+
+export const ADVANCED_PACKAGE_TEMPLATE_CONFIGS: PackageTemplateConfig[] = []
