+++ frontends/nextjs/src/lib/nerd-mode-ide/templates/configs/experimental.ts
@@ -1,0 +1,3 @@
+import type { PackageTemplateConfig } from '../../types'
+
+export const EXPERIMENTAL_PACKAGE_TEMPLATE_CONFIGS: PackageTemplateConfig[] = []
